feat: add integration test step to copy CA certs into container
This step will check for a filepath to a tar archive containing CA file(s) in the TEST_GRAFANA_CLOUD_CA_TAR_PATH env var. If a file has been provided, it will be copied to the vault container and the CA certs updated. This allows integration tests to be run when local CA certs are required to verify requests

Co-authored-by: Sam Owens <sam.owens@form3.tech>